Feat/table refactor (#84)
* move DocsTable into Table

* add Table docs
