+++ packages/styleguide/src/components/Table/Table.js
@@ -13,7 +13,9 @@ export const propTypes = {
   ),
   /** Data to print out in the table. */
   data: arrayOf(object),
+  /** Outside border */
   hasOutsideBorder: bool,
+  /** Responsivness of table */
   isResponsive: bool
 };
 
@@ -30,7 +32,9 @@ const Table = ({ columns = [], data = [], isResponsive, ...other }) => {
   const tableHead = columns.every(noLabel) ? null : (
     <thead>
       <tr>
-        {columns.map((column, i) => <th key={i.toString()}>{column.label}</th>)}
+        {columns.map((column, i) => (
+          <th key={i.toString()}>{column.label}</th>
+        ))}
       </tr>
     </thead>
   );
@@ -41,13 +45,12 @@ const Table = ({ columns = [], data = [], isResponsive, ...other }) => {
       <tbody>
         {data.map((row, i) => (
           <tr key={i.toString()}>
-            {columns.map(
-              column =>
-                column.render ? (
-                  column.render(row[column.key], column)
-                ) : (
-                  <td key={column.key}>{row[column.key]}</td>
-                )
+            {columns.map(column =>
+              column.render ? (
+                column.render(row[column.key], column)
+              ) : (
+                <td key={column.key}>{row[column.key]}</td>
+              )
             )}
           </tr>
         ))}
